pyproject.toml
[tool.poetry]
name = "altariq"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
uvicorn = "^0.30.5"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
postgres = "^4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
